+++ CHANGES.md
@@ -1,4 +1,8 @@
 
+v0.3.0
+
+- Add support for property replacement in the get method e.g. $config->get('db.${db.default}');
+- Add support for relative property replacement in the get method e.g. $config->get('db.${.default}');
 
 v0.2.0